Rename GitHub interaction to GitHub interactor
