--- modules/validation-pipeline/github-interaction/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubinteraction/GitHubCIJobProjectMetadata.kt
@@ -1,11 +1,0 @@
-package org.cafejojo.schaapi.validationpipeline.githubinteraction
-
-import org.cafejojo.schaapi.validationpipeline.CIJobProjectMetadata
-
-/**
- * Metadata of a GitHub project for a specific CI job run.
- */
-data class GitHubCIJobProjectMetadata(val checkRunId: Int, val installationId: Int, val fullName: String) :
-    CIJobProjectMetadata {
-    override fun getIdentifier() = checkRunId.toString()
-}
