--- modules/validation-pipeline/github-interaction/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubinteraction/githubapi/FuelConfiguration.kt
@@ -1,12 +1,0 @@
-package org.cafejojo.schaapi.validationpipeline.githubinteraction.githubapi
-
-import com.github.kittinunf.fuel.core.Client
-import com.github.kittinunf.fuel.core.FuelManager
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-internal open class FuelConfiguration(httpClient: Client) {
-    init {
-        FuelManager.instance.client = httpClient
-    }
-}
