--- modules/validation-pipeline/github-interaction/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubinteraction/webhookevents/CheckSuiteEvent.kt
@@ -1,12 +1,0 @@
-package org.cafejojo.schaapi.validationpipeline.githubinteraction.webhookevents
-
-internal data class CheckSuiteEvent(
-    val installation: Installation,
-    val action: String = "",
-    val checkSuite: CheckSuite,
-    val repository: Repository
-) {
-    fun isRequested() = action == "requested"
-}
-
-internal data class CheckSuite(val id: Int, val headBranch: String = "", val headSha: String = "")
