--- modules/validation-pipeline/github-interaction/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubinteraction/webhookevents/InstallationEvent.kt
@@ -1,10 +1,0 @@
-package org.cafejojo.schaapi.validationpipeline.githubinteraction.webhookevents
-
-internal data class InstallationEvent(
-    val installation: Installation,
-    val action: String = "",
-    val repositories: List<Repository>? = null
-) {
-    fun isCreated() = action == "created"
-    fun isDeleted() = action == "deleted"
-}
