+++ modules/validation-pipeline/github-interactor/gradle.properties
