+++ modules/validation-pipeline/github-interactor/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubinteractor/GitHubCIJobProjectMetadata.kt
@@ -1,0 +1,11 @@
+package org.cafejojo.schaapi.validationpipeline.githubinteractor
+
+import org.cafejojo.schaapi.validationpipeline.CIJobProjectMetadata
+
+/**
+ * Metadata of a GitHub project for a specific CI job run.
+ */
+data class GitHubCIJobProjectMetadata(val checkRunId: Int, val installationId: Int, val fullName: String) :
+    CIJobProjectMetadata {
+    override fun getIdentifier() = checkRunId.toString()
+}
