+++ modules/validation-pipeline/github-interactor/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubinteractor/webhookevents/CheckSuiteEvent.kt
@@ -1,0 +1,12 @@
+package org.cafejojo.schaapi.validationpipeline.githubinteractor.webhookevents
+
+internal data class CheckSuiteEvent(
+    val installation: Installation,
+    val action: String = "",
+    val checkSuite: CheckSuite,
+    val repository: Repository
+) {
+    fun isRequested() = action == "requested"
+}
+
+internal data class CheckSuite(val id: Int, val headBranch: String = "", val headSha: String = "")
