+++ modules/validation-pipeline/github-interactor/src/test/resources/fixtures/github/installation_deleted_webhook.resp
@@ -1,0 +1,63 @@
+{
+  "action": "deleted",
+  "installation": {
+    "id": 203592,
+    "account": {
+      "login": "cabotest",
+      "id": 40028399,
+      "node_id": "MDQ6VXNlcjQwMDI4Mzk5",
+      "avatar_url": "https://avatars0.githubusercontent.com/u/40028399?v=4",
+      "gravatar_id": "",
+      "url": "https://api.github.com/users/cabotest",
+      "html_url": "https://github.com/cabotest",
+      "followers_url": "https://api.github.com/users/cabotest/followers",
+      "following_url": "https://api.github.com/users/cabotest/following{/other_user}",
+      "gists_url": "https://api.github.com/users/cabotest/gists{/gist_id}",
+      "starred_url": "https://api.github.com/users/cabotest/starred{/owner}{/repo}",
+      "subscriptions_url": "https://api.github.com/users/cabotest/subscriptions",
+      "organizations_url": "https://api.github.com/users/cabotest/orgs",
+      "repos_url": "https://api.github.com/users/cabotest/repos",
+      "events_url": "https://api.github.com/users/cabotest/events{/privacy}",
+      "received_events_url": "https://api.github.com/users/cabotest/received_events",
+      "type": "User",
+      "site_admin": false
+    },
+    "repository_selection": "selected",
+    "access_tokens_url": "https://api.github.com/installations/203592/access_tokens",
+    "repositories_url": "https://api.github.com/installation/repositories",
+    "html_url": "https://github.com/settings/installations/203592",
+    "app_id": 12757,
+    "target_id": 40028399,
+    "target_type": "User",
+    "permissions": {
+      "checks": "write",
+      "metadata": "read"
+    },
+    "events": [
+      "check_run"
+    ],
+    "created_at": "2018-06-07T12:39:46.000Z",
+    "updated_at": "2018-06-07T12:39:46.000Z",
+    "single_file_name": null
+  },
+  "sender": {
+    "login": "cabotest",
+    "id": 40028399,
+    "node_id": "MDQ6VXNlcjQwMDI4Mzk5",
+    "avatar_url": "https://avatars0.githubusercontent.com/u/40028399?v=4",
+    "gravatar_id": "",
+    "url": "https://api.github.com/users/cabotest",
+    "html_url": "https://github.com/cabotest",
+    "followers_url": "https://api.github.com/users/cabotest/followers",
+    "following_url": "https://api.github.com/users/cabotest/following{/other_user}",
+    "gists_url": "https://api.github.com/users/cabotest/gists{/gist_id}",
+    "starred_url": "https://api.github.com/users/cabotest/starred{/owner}{/repo}",
+    "subscriptions_url": "https://api.github.com/users/cabotest/subscriptions",
+    "organizations_url": "https://api.github.com/users/cabotest/orgs",
+    "repos_url": "https://api.github.com/users/cabotest/repos",
+    "events_url": "https://api.github.com/users/cabotest/events{/privacy}",
+    "received_events_url": "https://api.github.com/users/cabotest/received_events",
+    "type": "User",
+    "site_admin": false
+  }
+}
