+++ modules/validation-pipeline/github-interactor/src/test/resources/githubtestreporter.properties
@@ -1,0 +1,3 @@
+app_id                   = 12345
+app_private_key_location = pkcs8_key
+tests_storage_location   = output/
