+++ modules/web/build.gradle
@@ -12,7 +12,7 @@ apply plugin: "org.springframework.boot"
 apply plugin: "io.spring.dependency-management"
 
 dependencies {
-    compile project(":validation-pipeline:github-interaction")
+    compile project(":validation-pipeline:github-interactor")
     compile project(":validation-pipeline:ci-job")
 
     compile group: "com.fasterxml.jackson.module", name: "jackson-module-kotlin"
