+++ settings.gradle
@@ -30,7 +30,7 @@ includeModule "mining-pipeline", "ccspan-pattern-detector"
 
 includeModule "validation-pipeline"
 includeModule "validation-pipeline", "ci-job"
-includeModule "validation-pipeline", "github-interaction"
+includeModule "validation-pipeline", "github-interactor"
 includeModule "validation-pipeline", "junit-test-runner"
 
 includeModule "web"
